Plans ページにおけるスワイプ動作のコードを綺麗に

+++ app/src/main/java/com/example/karaoke_note/Plans.kt
@@ -1,25 +1,37 @@
 package com.example.karaoke_note
 
 //noinspection UsingMaterialAndMaterial3Libraries
-//noinspection UsingMaterialAndMaterial3Libraries
-//noinspection UsingMaterialAndMaterial3Libraries
+import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material.ExperimentalMaterialApi
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.Delete
+import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Icon
 import androidx.compose.material3.ListItem
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.SnackbarDuration
 import androidx.compose.material3.SnackbarHostState
+import androidx.compose.material3.SnackbarResult
+import androidx.compose.material3.SwipeToDismissBox
+import androidx.compose.material3.SwipeToDismissBoxValue
 import androidx.compose.material3.Text
+import androidx.compose.material3.rememberSwipeToDismissBoxState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
@@ -30,7 +42,6 @@ import com.example.karaoke_note.data.Song
 import com.example.karaoke_note.data.SongDao
 import com.example.karaoke_note.data.SongScore
 import com.example.karaoke_note.data.SongScoreDao
-import com.example.karaoke_note.ui.component.DraggableBox
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import java.time.LocalDate
@@ -72,29 +83,35 @@ private fun removePlansListItem(
     songDao: SongDao,
     songScoreDao: SongScoreDao,
     scope: CoroutineScope,
-    snackBarHostState: SnackbarHostState
+    snackBarHostState: SnackbarHostState,
+    onUndo: () -> Unit
 ) {
     scope.launch {
         val snackbarResult = snackBarHostState.showSnackbar(
-            message = "One data has been deleted.",
-            actionLabel = null,
+            message = "One data has been removed",
+            actionLabel = "Undo",
             withDismissAction = true,
             duration = SnackbarDuration.Long
         )
-        // Undo 機能を実装したい
+        // スナックバーの Undo がタップされなかった場合にのみ削除処理を実行する
+        if (snackbarResult != SnackbarResult.ActionPerformed) {
+            // スコア ID を削除
+            //   これだけでは一度 Plans に登録した SongID や ArtistID などは残るので、サジェスト機能で出てくる
+            songScoreDao.deleteSongScore(scoreId)
+
+            // SongID の削除
+            deleteSongData(songId, songScoreDao, songDao, scope)
+            // ArtistID の削除
+            deleteArtistData(artistId, artistDao, songDao, scope)
+        }
+        else {
+            // Undo が押された場合は onUndo コールバックを実行する
+            onUndo()
+        }
     }
-
-    // スコア ID を削除
-    //   これだけでは一度 Plans に登録した SongID や ArtistID などは残るので、サジェスト機能で出てくる
-    songScoreDao.deleteSongScore(scoreId)
-
-    // SongID の削除
-    deleteSongData(songId, songScoreDao, songDao, scope)
-    // ArtistID の削除
-    deleteArtistData(artistId, artistDao, songDao, scope)
 }
 
-@ExperimentalMaterialApi
+@ExperimentalMaterial3Api
 @Composable
 fun PlansPage(
     songDao: SongDao,
@@ -120,71 +137,74 @@ fun PlansPage(
                     if (song != null) {
                         val artist = artistDao.getNameById(song.artistId)
                         if (artist != null) {
-                            /*
-                            val dismissState = rememberDismissState(
-                                confirmStateChange = {
-                                    if (songScoreList.isNotEmpty() && it == DismissValue.DismissedToStart) {
-                                        // Plan データを削除する
-                                        removePlansListItem(
-                                            artistId = song.artistId,
-                                            songId = songScore.songId,
-                                            scoreId = songScore.id,
-                                            artistDao = artistDao,
-                                            songDao = songDao,
-                                            songScoreDao = songScoreDao,
-                                            scope = scope,
-                                            snackBarHostState = snackBarHostState
-                                        )
-                                        true    // スワイプして songScore が消える
-                                    }
-                                    else {
-                                        false   // スワイプして元に戻る
-                                    }
-                                }
-                            )
+                            val isDeleted = remember { mutableStateOf(false) }
 
-                             */
+                            if (!isDeleted.value) {
+                                // SwipeToDismissBox の状態を管理する State
+                                val dismissState = rememberSwipeToDismissBoxState(
+                                    initialValue = SwipeToDismissBoxValue.Settled,
+                                    // スワイプが確定する距離的閾値を調整する
+                                    // 本当なら速度的閾値も調整したいのだが・・・
+                                    positionalThreshold = { it * 0.75f },
+                                )
+                                SwipeToDismissBox(
+                                    state = dismissState,
+                                    backgroundContent = {
+                                        val color = when (dismissState.targetValue) {
+                                            SwipeToDismissBoxValue.EndToStart -> MaterialTheme.colorScheme.errorContainer
+                                            else -> MaterialTheme.colorScheme.background
+                                        }
+                                        Row(
+                                            modifier = Modifier
+                                                .fillMaxSize()
+                                                .background(color)
+                                                .padding(horizontal = 24.dp),
+                                            horizontalArrangement = Arrangement.End,
+                                            verticalAlignment = Alignment.CenterVertically
+                                        ) {
+                                            Icon(
+                                                imageVector = Icons.Outlined.Delete,
+                                                contentDescription = "Delete",
+                                                tint = MaterialTheme.colorScheme.onErrorContainer,
+                                                modifier = Modifier
+                                            )
+                                        }
+                                    },
+                                    // 左から右へのスワイプを無効化する
+                                    enableDismissFromStartToEnd = false,
+                                    onDismiss = { dismissValue ->
+                                        // 右から左へのスワイプが完了した場合
+                                        if (dismissValue == SwipeToDismissBoxValue.EndToStart) {
+                                            // UI から即座にアイテムを削除する
+                                            isDeleted.value = true
 
-                            DraggableBox(
-                                modifier = Modifier,
-                                onDelete = {
-                                    if (songScoreList.isNotEmpty()) {
-                                        // Plan データを削除する
-                                        removePlansListItem(
-                                            artistId = song.artistId,
-                                            songId = songScore.songId,
-                                            scoreId = songScore.id,
-                                            artistDao = artistDao,
-                                            songDao = songDao,
-                                            songScoreDao = songScoreDao,
-                                            scope = scope,
-                                            snackBarHostState = snackBarHostState
-                                        )
-                                        true    // スワイプして songScore が消える
+                                            // Plan データを削除する処理を非同期で実行する
+                                            removePlansListItem(
+                                                artistId = song.artistId,
+                                                songId = songScore.songId,
+                                                scoreId = songScore.id,
+                                                artistDao = artistDao,
+                                                songDao = songDao,
+                                                songScoreDao = songScoreDao,
+                                                scope = scope,
+                                                snackBarHostState = snackBarHostState
+                                            ) {
+                                                // Undo が押されたら UI に再表示する
+                                                isDeleted.value = false
+                                            }
+                                        }
                                     }
-                                    else {
-                                        false   // スワイプして元に戻る
-                                    }
+                                ) {
+                                    // 表示するメインのコンテンツ
+                                    PlansListItem(
+                                        song,
+                                        songScore,
+                                        artist,
+                                        showEntrySheetDialog,
+                                        editingSongScore
+                                    )
                                 }
-                            ) {
-                                PlansListItem(song, songScore, artist, showEntrySheetDialog, editingSongScore)
                             }
-
-
-                            /*
-                            CustomSwipeToDismiss(
-                                state = dismissState,
-                                modifier = Modifier,
-                                directions = setOf(DismissDirection.EndToStart),
-                                background = {
-                                    BackGroundItem()
-                                },
-                                dismissContent = {
-                                    PlansListItem(song, songScore, artist, showEntrySheetDialog, editingSongScore)
-                                }
-                            )
-
-                             */
                         }
                     }
                     else {
@@ -195,32 +215,6 @@ fun PlansPage(
         }
     }
 }
-/*
-// SwipeToDismissBox の背面に表示されるコンテンツ (ゴミ箱)
-@Composable
-fun BackGroundItem() {
-    Column {
-        Box(
-            modifier = Modifier
-                .padding()
-                .fillMaxWidth()
-                .height(80.dp)
-                .background(MaterialTheme.colorScheme.errorContainer),
-            contentAlignment = Alignment.CenterEnd
-        ) {
-            Icon(
-                imageVector = Icons.Outlined.Delete,
-                contentDescription = "delete",
-                tint = MaterialTheme.colorScheme.error,
-                modifier = Modifier
-                    .size(60.dp)
-                    .padding(end = 30.dp)
-            )
-        }
-        HorizontalDivider(thickness = 1.dp, color = MaterialTheme.colorScheme.errorContainer)
-    }
-}
-*/
 
 @Composable
 fun PlansListItem(
